Bug in interconnect
IF a cache evicted a line, it would notify coherence.
Coherence could generate a DATA request for the bus if necessary.
The interconnect was sending *all* pending requests to every coherence /
cache, but this meant all other caches that had an outstanding request
for that address thought they just received DATA.